Specify -O3 for func attribute override
Starting with libgcrypt 1.9.0, the src/hwf-x86.c source file defines a
macro named FORCE_FUNC_FRAME_POINTER that (for gcc >= 4.7.0) expands to
a function attribute to override optimization options. This attribute is
applied to a function, `is_cpuid_available()`, defined a few lines below
the macro definition. The function attribute specifies a -f optimization
option, "no-omit-frame-pointer", thus applying the
`-fno-omit-frame-pointer` optimization specifically for the
is_cpuid_available() function.

However, gcc is failing to compile this source file for the 32-bit
build, raising an ICE. The file is compiled successfully for 64-bit.

I discovered a workaround by prepending an optimization *level* option
for the function attribute. Specifically, I prepended `-O3`, the current
option passed to gcc on the command line for libgcrypt 32-bit and 64-bit
builds. Note that the syntax for setting that option in the function
attribute is "O3".

The root cause is unknown. The Clear Linux OS `gcc` package carries
several patches that tweak optimization defaults, so perhaps they
influence the behavior of specifying optimization overrides via function
attributes.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>